+++ apps/web/templates/application.html.slim
@@ -23,7 +23,7 @@ html lang='en'
                 li.nav-item
                   = link_to 'Записать рецепт', routes.new_profile_recipe_path, class: 'nav-link btn'
                 li.nav-item
-                  = link_to 'Списки покупок', Web.routes.path(:shopping_lists)
+                  = link_to 'Списки покупок', routes.profile_shopping_lists_path, class: 'nav-link btn'
                 li.nav-item
                   = form_for :session, Web.routes.path(:session), method: 'delete' do
                     - submit 'Выйти', class: 'nav-link btn'
