+++ apps/web/views/profile/shopping_lists/index.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Web
+  module Views
+    module Profile
+      module ShoppingLists
+        class Index
+          include Web::View
+        end
+      end
+    end
+  end
+end
