+++ apps/web/views/profile/shopping_lists/show.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Web
+  module Views
+    module Profile
+      module ShoppingLists
+        class Show
+          include Web::View
+        end
+      end
+    end
+  end
+end
