--- apps/web/views/shopping_lists/show.rb
@@ -1,11 +1,0 @@
-# frozen_string_literal: true
-
-module Web
-  module Views
-    module ShoppingLists
-      class Show
-        include Web::View
-      end
-    end
-  end
-end
